Refactor admin controllers (#121)
* Split AdminController

* Move HolidayAdminController into Admin directory

* Rename controller methods

* Make auth check as a middleware

* Move the admin part of PressController into Admin directory

* Move admin twig files into admin directory

* Move admin part of UsersController into Admin directory
