+++ assets/views/admin/user/image.twig
@@ -47,7 +47,7 @@
 
       $('#select-btn').click(function () {
         var uid = $('#uid').val();
-        window.location.href = '/users/image_upload/' + uid;
+        window.location.href = `/admin/user/${uid}/image`;
       });
 
     });
