+++ src/Intra/Service/Menu/MenuService.php
@@ -93,11 +93,11 @@ class MenuService
                 new LinkList('관리자', [
                     new Link('권한 설정', '/admin/policy', new OnlyPolicyRecipientEditable()),
                     new Link('메일 수신 설정', '/admin/recipient', new OnlyPolicyRecipientEditable()),
-                    new Link('직원 목록', '/users/list', new OnlyUserManager()),
-                    new Link('휴가 조정', '/holidayadmin/', new OnlyHolidayEditable()),
+                    new Link('직원 목록', '/admin/user', new OnlyUserManager()),
+                    new Link('휴가 조정', '/admin/holiday', new OnlyHolidayEditable()),
                     new Link('회의실 설정', '/admin/room', new OnlyPolicyRecipientEditable()),
                     new Link('회의실 정기 예약', '/admin/event_group', new OnlyPolicyRecipientEditable()),
-                    new Link('보도자료 관리', '/press/', new OnlyPressManager()),
+                    new Link('보도자료 관리', '/admin/press', new OnlyPressManager()),
                 ], 'wrench'),
                 new Link('내정보', '/users/me', new PublicAuth(), null, 'user'),
                 new Link('로그아웃', '/usersession/logout', new PublicAuth(), null, 'log-out'),
